Extract start action
